pepc: do print a warning for the 'turbo' option
The '--turbo' option is always applied to the entire system (all CPUs). And we
have a check for this, but it did not work for the case when '--turbo' was used
without an argument, and this patch fixes the issue.

Before (no warning):

$ pepc pstates config --turbo --cpus 10
Turbo support: yes
Turbo enabled: yes

After (warning printed):

$ pepc pstates config --turbo --cpus 10
pepc: warning: the turbo setting is global, '--cpus', '--cores', and '--packages' options are ignored
CPU number: 10
Turbo support: yes
Turbo enabled: yes

Also change the warning message, because the old warning does not work for the
'--turbo' with no argument case.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>