pepc: move 'cpuidle' and 'cpufreq' objects creation up
This patch has no functional changes, pure refactoring. All it does is:

* Add 'cpuidle' argument to 'handle_cstate_config_options()'.
* Add 'cpufreq' argument to 'handle_pstate_config_options()'.

This improves readability a bit because these functions become a bit smaller and
simpler. Their calling functions are small, and it is OK to make them a bit
large.

The 'cpuidle' change will be helpful soon, the 'cpufreq' change was made just
to keep things consistent.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>