pepc: fix scope name
The "cpu" name is wrong for the scope, it should be "CPU. While this is not a
bug in this function, it confused me, because I saw "CPU" in other places. Fix
this and let's be more consistent about it.

Remove the 'lower()' call - we'll catch incorrect scope names this way.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>