pepc: allow for C-state changes per-CPU despite the scope
Today 'pepc' cannot be used for changing package-scoped features for just a
subset of CPUs in the packet. For example, if one tries to change "C1E
autopromote" for just CPU0, 'pepc' prints a warning and changes 'C1E
autopromote' for entire package.

However, sometimes we know what we are doing, and for experimenting purposes we
do want to change package-scoped features to just some CPUs in the package.

This patch changes 'pepc' behavior. Now 'pepc' will print a warning, and
proceed changing the feature for the selected CPUs. Assuming the user knows
what he/she is doing.

This patch does not cover the "printing" part, that will come a bit later.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>