pepc: print package-scoped features for every CPU
We were printing package-scoped features only for the entire package, but not
really correctly, because we were assuming that we could read the feature value
from a single CPU, and it would be the same for all of them. But this is
assumption is incorrect.

Ideally, we should recognize the situation when package-scoped feature has
different values on different CPUs, and warn the user, but this would be work
for future. For now, just print all C-state features on per-CPU basis.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>