pepc: improve C-state features printing
There is what happens when running
$ pepc cstate config --c1-undemotion --c1-demotion

Before this patch.

We call 'CPUIdle.get_cstates_config()' for "c1-undemotion", then call it for
"c1-demotion". So end up running it 2 times. If the features come from the same
MSR, we make it hard for 'CPUIdle.get_cstates_config()' to optimize and read
the MSR only once.

After this patch.

We call 'CPUIdle.get_cstates_config()' once for both features. This way it has
a possibility to optimize potentially slow MSR access and read it only once.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>