pepc: rework C-state features printing some more
This patch streamlines the feature printing code and simplifies it by removing
most of the per-feature "if" statements.

This patch also changes the behavior of the tool. Before this patch, we tried
to print C-state names only once, not for every CPU. But this required special
handling and was complex, so I simplified it and now we print this information
for every CPU.

My reasoning for the behavior change was also that hybrid CPUs like AlderLake
have cores with different capabilities. So generally, different CPUs may have
different C-states, so it is better to keep everything per-CPU.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>