CPUIdle: improve handling of unsupported features
The 'CPUIdle.get_cstates_config()' method can be called for many keys, and if
any of the features is not supported, it currently raises and exception. This
is problematic for use cases like

$ pepc cstates config --cstate-prewake --c1e-autopromotion

If the platform does not support the C-state prewake feature, pepc just exits
with error, without printing the C1E autopromotion status.

This patch teaches 'CPUIdle.get_cstates_config()' to avid raising exceptions
for unsupported features, and instead, just set the "feature_name_supported"
key to "False".

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>