pepc: print information about unsupported features too
When pepc is used like this:

pepc cstates config --feature_name

it prints information about the 'feature_name' features when the platform
supports the feature. But if the feature is not supported, no information is
printed. With this patch, the tool will just print that the feature is not
supported.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>