CPUIdle: add keys description to 'props'
The 'CPUIdle.props' is a per-property dictionary, which already has the list of
keys associated with the property. However, it does not include description for
those keys, the description is in a separate dictionary, that we also expose to
the user.

This is a bit strange structure. Let's move the descriptions to
'CPUIdle.props' instead, and unimport the separate keys description dictionary.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>