_FeaturedMSR: introduce 'get_feature()'
Introduce a 'get_feature()' method, similar to 'set_feature()'. This method
does the same as (now removed) 'feature_enabled()' method fore boolean
features.

For non-bolean features, it calls the special handler, which the sub-class must
implement. The only feature that needs this is the "pkg_cstate_limit" feature,
and the special method is not implemented for it yet.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>