CPUInfo: rename and adjust 'get_package_list'
Functions like this are called 'normalize_something' in other modules, let's be
consistent and call it 'normalize_packages()'. Also, let's not sort the
normalized list and leave the original order. We do not know if the user really
want/need the sorting, so it is safer not to change the order.

Adjust all the user of this funciton too.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>