PCStateConfigCtl: rework pkg C-state limit read
Implement the '_get_pkg_cstate_limit()' function, which will be called from
'get_feature()'. Make this function return the 'pkg_cstate_limit' feature
information dictionary.

Since 'get_feature()' accepts only one CPU, simplify the implementation and
support only one CPU.

Remove the "smartness" where we tried to read the limit on all CPUs and select
the strictest one - this complexity is not needed, it is based on assumptions
that are not necessarily correct too. Just read it for the requested CPU and
return whatever we read. Let upper levels deal with the rest.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>