CPUFreq: close only objects we created
Make sure close/destroy all the objects that we created, and we do not
close/destroy any object that we did not create (user provided them).

Add the 'lscpu_info' optional argument - users may have it and it is OK to let
them pass it to the constructor.

Improve and rename the '_check_uncore_freq_supported()' function. Move the
uncore driver loading to it. This way, we don't have to load the driver unless
we do uncore operations.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>