CPUIdle: improvement: create less objects
In 'CPUIdle' we use 'PCStateConfigCtl' and 'PowerCtl', both of which create
'lscpu_info' and 'msr' object, unless those are provided.

If the user already has an 'lscpu_info', let him/her pass it to 'CPUIdle'. This
will make sure that  'PCStateConfigCtl' and 'PowerCtl' won't create 2
additional instances of 'lscpu_info'.

If user did not provide 'lscpu_info' or 'msr', create one instance in
'CPUIdle', and pass it down to 'PCStateConfigCtl' and 'PowerCtl'. Otherwise
they will create 2 instances of these objects.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>