CPUIdle: add 'get_props()' method
Add a 'get_props()' method which can be used for getting property values. This
method is consistent with 'set_prop()', but it retrieves values of properties.

The method receives the list of properties to read, and the list of CPUs to
read them for, and yields dictionaries for every CPU. The dictionary describes
all the requested properties.

This method is similar to 'get_cstates_config()', but works with properties,
not with keys. Note, each property may have multiple keys.

The 'get_cstates_config()' method will be removed soon.

Also add the 'pkg_cstate_limit_aliases' key support.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>