PCStateConfigCtl: prepae for fixing pkg_cstate_limit bits
The package C-state limit bits are (3,0) on some platforms, and (2,0) on other
platforms. But our code assumes it is always (3,0). Refactor the code to stop
assuming the same bits for all CPU models.

The upcoming pathces will adjust the bits for different platroms. This patch is
just a preparation that makes the "bits" changable.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>