PCStateConfigCtl: add "locked" feature
Add the lock bit as a separate feature. The follow up patches will remove it
from the "pkg_cstate_limit" feature, where it is coupled to right now. This is
just a more consistent and uniform way of handling MSR fields.

In CPUIdle, however, we do not want "locked" to be a separate property, we want
it to be part of the "pkg_cstate_limit" property. So we have to remove it from
'PROPS'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>