MSR: remove 'toggle_bit()'
The problem I had with this method is that it is confusing in the follwoing
way. The name assumes that it will set a bit to 0 if it was 1, and sets the bit
to 1 if it was 0. Therefore, this function should not receive the bit value at
all. However, it did, and this was very confusing.

This function is not used anywhere now, so we can just remove it.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>