MSRs: use "on" / "off" instead of "enabled" / "disabled"
For boolean options, let's use "on" / "off" values, instead of "enabled" /
"disabled". At this point there is no fundamental difference, but since we use
"on" / "off" in command line and prints, let's be consistent.

But a bit longer term, this will allow us to streamline booleand and integer
options handling.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>