_FeaguredMSR: hide the 'supported' flag
I foresee that in the future we may have to deal with hybrid CPUs where MSR
features may be supported only on a subset of CPUs. However, currently we
expose a global "supported" flag via the 'features' dictionary. Users should
not use it, and they should use the 'feature_supported()' method instead.

This patch removes the 'supported' flag from the 'features' dictionary, so that
users can't even access it. Instead, the flag is stored in a separate, private
'_features' dictionary.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>