pepclibs: fix 'close()' methods in several modules
In the 'close()' method we are trying to close objects like 'proc' only if we
created them, as opposed to receiving them from the caller.

We use attributes like 'self._close_proc' to track whether the objects must be
closed (because we created them) or not.

However, due to a bug we were checking wrong attributes, for example
'self._close__proc' instead of 'self._close_proc'. This patch fixes the
problem.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>