CPUInfo: do more in '_get_lscpu()'
Currently '_get_lscpu()' runs 'lscpu' and stores its output in
'self._lscpu_cache' basically as it is.

Then in 2 places we process the cached output and parse it. And there is some
code duplication: ignoring commentaries and splitting the lines.

Move these common steps to '_get_lscpu()'. This removes some of the code
duplication.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>