CStates: simplify the 'get_props()' dictionary
This is the first patch in an effort to simplify the dictionary returned by
'get_props()'. The end goal is to make it flat and get rid of the "keys" stuff.

In this patch is based on an observation that 'get_props()' dictionary contains
the main property value twice: in the "val" key, and in "keys" -> "pname" key.
This patch removes the "val" key.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>