CStates: change 'get_cstates_info()' API
Make the dictionaries that 'get_cstates_info()' yields look very similar to
what 'get_props()' yields.

Before this change, 'get_cstates_info()' yielded a dictionary for every
C-state. Now it yields a dictionary for every CPU, and the dictionary includes
all the C-states. This is simplar to 'get_props()', which yields per-CPU
dictionaries.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>