_PepcCStates: improve cstates info output
Currently, the command prints information about every C-state on every CPU.
This is a lot of output and it is not so useful.

This patch removes some of the output, like time and and usage. Then it
aggregates all the output by CPUs, and prints a concise summary. Example:

$ pepc cstates config --enable all --cpus all
$ pepc cstates config --disable C1E --cpus 1,5,6
$ pepc cstates info --cpus all

POLL: 'off' for CPUs 0-143 (all CPUs)
       - expected latency: '0'
       - target residency: '0'
C1: 'off' for CPUs 0-143 (all CPUs)
     - expected latency: '1'
     - target residency: '1'
C1E: 'off' for CPUs 0,2-4,7-143
     'on' for CPUs 1,5,6
      - expected latency: '4'
      - target residency: '4'
C6: 'off' for CPUs 0-143 (all CPUs)
     - expected latency: '170'
     - target residency: '600'

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>