tests: make dataset option private
The 'dataset' option can be used to select which
emulated dataset is used for testing. To extend
the usage and dynamically search for all the
existing datasets we need to use the option in a
bit different way. For now, stop parametrizing
all the test functions with 'dataset' option and
define available datasets as fixture parameters.
This will be changed further in next patches.

Signed-off-by: Antti Laakso <antti.laakso@intel.com>