CStates: fix 'pepc cstates config' crash
This is a quick fix for a crash in Ubuntu which has the following symptoms:

$ pepc cstates config --enable all
pepc: error: the following command failed with exit code 127:
xargs -a '/tmp/_linuxcstates_tX7ZgZ/fpaths.txt' cat

/bin/sh: 1: exec: --: not found

This is related to the 'shell=True' option, which does not necessarily preserve
'PATH'. But for the specific command in question, shell=True is not needed, so
removing it. Longer term we need to have a fall-back method for toggling
C-states, which does not depend on 'xargs' and 'cat'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>