Procs: rename 'proc' to 'task'
We use the 'proc' variable names for a a running process (a task) in 'Proc.py',
while outside of 'Proc.py' we use 'proc' for 'Procs.Proc()' instances. This
naming is confusing and inconsistent. Renaming 'proc' inside 'Procs.Proc()' is
much easier than renaming 'proc' in all the other places, so this patch renames
'proc' to 'task' when it is used for a running process.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>