Procs/SSH: introduce 'Task' class
This patch itroduces a 'Task' class representing local and remote tasks. The
intention is to stop live-patching paramiko channel and Popen objects, and add
all custom fields to 'Task' objects instead. Additionally, this change will
help unifying common bits between local and remote tasks in a better way.

This patch is the smallest first step I could come up with. Many attributes
that have to be moved to 'Task' objects are still in the Channel/Popen ojects.
This will be done later.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>