PStates: load uncore freq driver only if necessary
Before this patch: the uncore frequency driver is always attempted to be loaded
by the PStates module.

After this patch: the uncore frequency driver will be attempted to be loaded
only if an uncore property is requested. This is more optimal - no need to load
the driver without a real reason.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>