ProcessManager: add more FS helper methods
Add the following methods to the process manager:
  * read()
  * write()
  * get_mtime()
  * which()
  * abspath()
  * mkdtemp()
  * rmtree()

These methods are similar to what we have in 'FSHelpers'.

The 'mmtemp()' was renamed to 'mkdtemp()' to be consistent with
'tempfile.mkdtemp()'.

The 'rm_minus_rf()' function was renamed to 'rmtree()' to be consistent with
'shutil.rmtree()'.

The 'default' argument was renamed, because it is ugly. Use 'must_exist'
instead.

The plan is to gradually move many 'FSHelpers' functions to the process
manager.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>