_Pepc: improve error message
When user does not provide the sub-command, pepc fails with a
"use '-h' for help" message, which is not very clear, and we got a report that
pepc does not work. Example:

$ pepc cstates
pepc: error: please, run 'pepc -h' for help

With this patch pepc will print something like:

$ pepc cstates
usage: pepc cstates [-h] [-q] [-d] {info,config} ...
pepc cstates: error: the following arguments are required: {info,config}
Use -h for help.

Which should be more descriptive.

The side-effect of this patch is a slightly messier help message, but this is
not a big deal. Example:

$ pepc cstates -h
usage: pepc cstates [-h] [-q] [-d] {info,config} ...

Various commands related to CPU C-states.

options:
  -h             Show this help message and exit.
  -q             Be quiet.
  -d             Print debugging information.

further sub-commands:
  {info,config}
    info         Get CPU C-states information.
    config       Configure C-states.

The "messy" part is the "{info,config}" part.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>