SSHProcessManager: fix timeout crash and increase timeout
When we start the process, we first read its PID, and we expect to get the PID
within 10 seconds. If we do not get the PID, we are supposed to error out.

However, because of a bug, we crash instead with the following symptom:

File "pepclibs/helperlibs/SSHProcessManager.py", line 217, in _wait_intsh
  self._handle_queue_item(streamid, data, capture_output=capture_output,
File "pepclibs/helperlibs/_ProcessManagerBase.py", line 148, in _handle_queue_item
  data, self._partial[streamid] = extract_full_lines(self._partial[streamid] + data)
TypeError: can only concatenate str (not "NoneType") to str

This patch should fix the problem (did not verify). It also increases the
timeout from 10 to 60 seconds.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>