PStates: rewrite 'set_props()' to use 'scope'
Rewrite 'set_props()' to use scope, which saves time.

TL;DR: All cpus for passed property are wiped. Then for each cpu we check if it
has the value cached, if not we write the value and cache it according to scope.
So the next cpu, if it belong to the same scope it will no be writen again.
Setting frequency may be tricky, because there are ordering constraints, so it
is done separately, so we validate, set and cache them separetly.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>