LocalProcessManager: improve exceptions
File operations like 'write' and 'close' may sometimes raise the
'PermissionError' exception. For example, 'close()' raises it when 'root'
writes something like 'performance' to the
'/sys/module/pcie_aspm/parameters/policy' file, and then closes the file. In
this case, 'write()' does not necessarily fails, but 'close()' does.

As a result, we raise the 'Error' type exception (because of the
'WrapExceptions' class). However, we want to translate it to the
'ErrorPermissionDenied' type instead.

This is exactly what the patch does - it wraps the 'close' and 'write' methods
so that they translate 'PermissionError' to 'ErrorPermissionDenied'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>