Human: fix negative numbers conversion
Fix '_tokenize()' to not fail on numbers like '-1' or '1.1'. The problem was
that 'isdigit()' returns 'False' on them, so we fail to append the default unit
and end up raising an Error. Use 'Trivial.is_num()' instead.

This change will fix all the users of '_tokenize()', for example,
'parse_duration_ns()'.

Before this patch, 'parse_duration_ns()' raised an error on '-1' and '1.1'.
After this patch, it correctly parses them and turns into nanoseconds.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>