pcstates_common: add 'get_fellows()'
'get_fellows()' creates a dict with keys being scopes and values being all
fellow CPUs (CPUs sharing the same scope as a CPU) to the CPU passed to the
function. The dict is used when running tests with function calls that require
a list of CPUs with the same scope as the CPU that we are changing the value of.

Example:
 - If we want to set property 'min_uncore_freq' with 'set_prop()' we need to
   pass a list of all CPUs in the same 'die' scope as the CPU which value we
   are changing.

This function will be used in the next commits.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>