_PepcCommon: do not fail on 'tuned' check
Checking for 'tuned' is just an optimization. Today, if 'systemctl' is not
installed, we fail. This is wrong, because for sure 'tuned' is not installed if
'systemctl' is not available. With this patch, we'll just ignore the 'systemctl
not found' error. And in case of any other error, let's just print a warning.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>