PCStates: rename a private method
Observation: '_set_cpu_prop_value()' and '_get_cpu_prop_value()' sound like
they are the same, but one for writing, and the other for reading.

However, they are different, and this naming is confusing. Indeed, we use "cpu"
in function name only when it is for a single CPU, so the
'_set_cpu_prop_value()' name is inconsistent because it works for many CPUs,
not just one.

Rename '_set_cpu_prop_value()' to '_set_prop_value()' for the sake of
consistency.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>