PStates: improve errors handling
If we fail to read a sysfs attribute, and if we also fail to get the property
with the "getter" function, we will end up with an exception that includes only
the latter failure details, but not the former.

This patch improves the situation. Now we raise an error with a message that
includes both the original error message (from the sysfs read failure) and the
"getter" function error message.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>