pepc: do not print unsupported options by default
When we print information about all the options, we also print about
unsupported ones, which adds clutter and in most cases - not needed. Let's stop
doing this and just skip unsupported options.

So, say, if EPP is not supported, then this command will not print about it:
$ pepc pstates info

But this command will print:
$ pepc pstates info --epp
Energy Performance Preference: not supported for CPUs 0-7 (all CPUs)

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>