PStates: never report 'intel_cpufreq' for the 'intel_pstate' driver
In the sysfs 'scaling_driver' file, the 'intel_pstate' driver is names
'intel_pstate' when it is in active mode, and it is names 'intel_cpufreq' when
it is in passive mode. So there are 2 names for the same kernel driver.

Since in general, users know what is 'intel_pstate' and they do not know what
is 'intel_cpufreq', let's always report driver name as 'intel_pstate'
regardless of the mode it is in.

So when 'intel_cpufreq' is read it is saved as 'intel_pstate'.

Artem: minor adjustments, commit message improvements.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>