ProcessManager: tweak 'rsync' method API
Make it possible to use the 'opts' argument for any sort of options, not
only single-letter options. For this, require that the user specifies
also the leading dash symbol. This way, user can also pass something
like "--exclude mydir -r".

While on it, swithch 'LocalProcessManager.rsync()' to f-strings, for
consistency with the rest of the code.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>