PStates: save CPU frequency writing function to variable
This patch is in preparation for making 'min_freq_hw' and 'max_freq_hw'
writable.

Properties 'min_freq_hw' and 'max_freq_hw' will write to MSR instead of sysfs,
and will need a different function. Changes made in this patch will remove the
need for multiple "if else" in the code to determine which write function to use.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>