PStates: fix and simplify 'max_turbo_freq' property
Before this patch, the 'max_turbo_freq' property fallen back to
'max_freq' property if we could not read it from an MSR. But this was
wrong, because it should have read 'max_freq_limit' instead.

This patch fixes the bug and simplifies 'max_turbo_freq' property by
simply turning it into MSR-only property. This is simpler and more
straight forward.

Generally, our properties should be of only 2 types:
1. MSR-only
2. Try OS knobs first, fall-back to MSR.

Having properties that start with an MSR and fall-back to OS makes the
logic harder to follow.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>