CPUInfo: fix die numbers
When a CPUs "die_id" file was not found, '_no_die_info' was set to 'True'. This
caused all subsequent CPUs die numbers to be 0, even if a sebsequent CPU had a
"die_id" file.

This patch fixes this issue, by making it so that all non cached die numbers
will try to read the "die_id" file.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>