Logging: add 'force_tty_newline_prefix' attribute
In some cases users print to 'stdout' directly, for example using
'print("\rProgress: %d done")'. In this case, the next logging message will
not start from the new line, which does not look pretty.

Introduce the 'force_tty_newline_prefix' attribut that will make sure that
any log message that ends up in a TTY will be prefixed with a newline.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>