pepc: address pylint warnings and use 'Error().indent()'
In many places we use a pattern like this:

  raise Error("blah blah blah: {err}")

which may result in a very long error message. Improve this by using the
following pattern:

  raise Error("blah blah blah:\n{err.indent(2)}")

which results in a more human-readable error message.

In addition to this, address all the pylint warnings. Mostly about missing
encoding in 'open()' calls, but there also were few other.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>