CStates: improve error message
When a system has no C-states because it was booted with 'idle=poll' kernel
boot parameter, 'pepc cstates info' fails, but the error message is not ideal.
It simply says that C-states were not found. Lets improve it and make it say
something like this:

pepc: error: failed to find C-state files in '/sys/devices/system/cpu'.
There is no idle driver in use, which may be why Linux C-states support was not found.
You have the 'idle=poll' kernel boot parameter, which may be why there is no idle driver.

This is just more helpful.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>