pepc cstates: make '--pkg-cst-limit' case-insensitive
This is a fix for a regression where a command like

$ pepc cstates info --pkg-cstate-limit pc0

fails complaning that "pc0" is not a valid name, and it should be "PC0"
instead. At some point in the past it work, but then broke.

We have all the code for handling the names in a case-insensitve way, but it
did not work because the code expected a non-existing "choice" type. The
'pkg_cst_limit' property is of "dict" type. This patch replaces "choice" with
"dict", which fixes the issue.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>