EmulProcessManager: add support for emulating '/sys/devices/system/cpu/online'
There are two types of CPU online files. One containing all online CPUs, in
number range format, e.g. "0-3,5,9-13,15", this file is updated automatically
when a CPU online state is modified. The other is per CPU containing "1" if
online and "0" if offline, and can be used to online/offline a CPU.

In this patch when tring to read '/sys/devices/system/cpu/online' on an emulated
system, the read call is redirected to '_online_read()'. This function will read
all CPU specific online files and return the online CPUs in number range format.
Thus, emulating the automatic update.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>