CPUInfo: remove offline CPUs from topology list
Sysfs does not provide any topology information for offline CPUs. Thus, offline
CPUs in CPUInfos topology have only "CPU" level value, for all other levels are
unknown. This makes it so that offline CPUs serve no other purpose than to state
that they are offline, which strictly speaking is not a topology.

In previous commits '_get_online_cpus()' and 'get_offline_cpus()', were modified
to not depend on CPUInfo topology for finding online or offline CPUs. This makes
it so that CPUInfo no longer needs to store offline CPUs or have an online flag
in internal topology structure.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>