PCStateConfigCtl: fix MSR_PKG_CST_CONFIG_CONTROL support
Our MSR_PKG_CST_CONFIG_CONTROL support is broken for the 'pkg_cstate_limit'
feature.

Problem: we assume that if we modifying PC-state for CPU A, and we read if from
CPU B, we'll read the same value as we wrote from CPU A. However, this is not
the case. But this is not the case at least on some Intel platforms. Here is an
example of how this is problematic.

Initial state: CPU A has limit PC6
               CPU B has limit PC6

We try to modify the limit to PC0.

We run 'pepc cstates config --pkg-cstates-limit PC0'. This ends up calling
'MSR().write(sname="package")'.

The MSR module will write PC0 limit only for CPU A. It will _not_ write for CPU
B. Logically this is correct, the MSR has package scope. And this will actually
work, the system will stop using PC6 and will use only PC0.

Now the problem is that when someone else reads the MSR from CPU B, he/she will
read PC6 limit. But from CPU A he/she will read the PC0 limit. The actual
limit is PC0, because this is the last-written value.

If someone runs, say, turbostat, it may decide to read the MSR from CPU B, and
report that current limit is PC6 (incorrect).

Consider if someone decides to change C1 demotion, which is in the same
MSR_PKG_CST_CONFIG_CONTROL as the PC-state limit, The person reads the MSR from
CPU B, flips the C1 demotion bit, and writes it back. And the person not only
will change the C1 demotion, but he/she will also change the current limit to
PC6.

Introduces notion of "write scope" - the scope to assume when writing the MSR.
In case of MSR_PKG_CST_CONFIG_CONTROL the changes are reflected on both logical
CPUs of a core, so the write scope name (wsname) is "core".

Also adjust all the 'MSR().write*()' callers to pass "wsname" when writing.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>