CPUInfo: add argument for selecting which CPUs to add level numbers for
This patch is purely an optimization.

There might be cases when we update topology only for a small fraction of
the total CPUs. Thus, it is better to specify which online CPUs we want to
add level numbers for, instead of reading topology for all online CPUs.

Note, NUMA node is the only exception. Because it does not store topology
info per-CPU in sysfs, so this optimization does not work in this case.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>