pepc topology: do not require --online-only
The following command fails:
$ pepc topology info --cpus all --core-siblings 1
pepc: error: '--core-siblings' must be used with '--online-only'

It is not intuitive why --core-siblings requires --online-only, and also too
much typing. The command can work just fine if a CPU is offline. I will just be
skipped, because we do not know its topology.

This is also inconsistent with other commands, like 'pepc cstates info' - it
does not have '--online-only'.

Remove this requirement.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>