CPUOnline: re-work how CPUOnline uses CPUInfo
Before this patch, each time CPUOnline 'online()' or 'offline()' methods
were called, CPUInfo object was re-created. In this patch instead of
re-creating the object we call 'cpus_hotplugged()'. This function will
notify 'CPUInfo' that the topology and online/offline CPU list must be
updated the next time they are needed.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>