HWPRequest: improve 'hwp' flag validation
To access 'MSR_HWP_REQUEST', bit 0 in 'MSR_PM_ENABLE' has to be set. The
bit has package scope, so we read the MSR for one CPU per package and set
the feature to 'supported=False' if the bit is 0.

In this patch instead doing this per feature, we read the CPUs once and
the apply the results to all features. This reduces 'read_cpu_bits()' calls
from 'packages * features' to 'packages'.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>