CPUInfo: improve 'normalize_cpus()'
In this patch we switch 'get_cpus()' to '_get_online_cpus()' which is faster.
'_get_online_cpus()' returns a pointer to the internal online CPU list which is
fine as long as 'cpus' is not "all", because in this case the list is only used
inside 'normalize_cpus()'. But when 'cpus' is "all", we return a copy of the
list, this also fixes the case when 'offlined_ok=True', because then
'_get_all_cpus()' is called which also returns a pointer to a list.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>