ArgsParse: optimize 'parse_int_list()'
'parse_int_list()' takes string of numbers separated by a comma, a list of
'ints' or just an 'int'/'str' and returns them as a list. The function in its
current state is only optimized for strings. In this patch the functions
is modified to not convert 'int' to string unnecessarily.

Argument                                        Faster
"0"                    				~137%
"0-300"            				~245%
0  						~279%
range(0, 400)					~844%
",".join(str(x) for x in range(0, 400))		~192%

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>