CPUInfo: save online/all CPUs as 'set' instead of 'list'
========================= Prolog ==========================
One of the most commonly used function is Pepc is 'normalize_cpus()'. This
function validates CPU numbers and returns a normalized list. Due to how
often it is called (e.g, 23762 times 'pepc pstates info -D spr1'), it is
imperative that the function is as fast as possible. Previously, I thought
that the slowest part of the function was comparing each CPU to a set of
valid CPUs, that is not the case. The slowest part is converting valid CPU
list to a 'set', taking ~65% of 'normalize_cpus()' runtime.

To improve 'normalize_cpus()' performance, the internal cached CPU lists are
now sets. This reduces 'normalize_cpus()' total footprint from 0.237s to
0.081s when running: 'pepc pstates info -D spr1'

======================== FAQ ===============================
 - How does this impact the rest if the code?
In all cases there are minor improvements, except for 'get_cpus(order="CPU")'.
The time footprint for 'get_cpus(order="CPU")' is ~750% larger. This
may seem like an issue, but the function is called only 58 times, and the time
increase is from 0.00000095 to 0.00000715. Compared to all the other
functions time decrease, this time increase is insignificant.

======================= TL:DR ==============================
Saving online and all CPUs as a set improves 'normalize_cpus()' performance by
~65%.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>