CPUInfo: remove support for CPU comma-separated string
Note: this commit only modifies CPUInfo.py, I am aware that other modules
docstrings have to be adjusted. To simplify this commit I have moved the
docstring changes to the next commit.

Changes:
 - Functions in CPUInfo.py no longer support comma-separated string of
   CPUs, packages, cores etc.

 - Functions that have 'cpus' as argument require them to be in a iterable,
   and if the function returns CPUs it will be list. For example:
   	'normalize_cpus(cpus=[1,2,3,4])' -> [1,2,3,4]
  	'normalize_cpus(cpus=[1])'       -> [1]
  	'normalize_cpus(cpus=1)'         -> Error

 - Functions that have 'cpu' as argument require it to be an 'int', and
   if the function returns a CPU it will be an 'int'.
  	'normalize_cpu(cpu=1)'           -> 1
  	'normalize_cpu(cpu=[1])'         -> Error

We have to specifically check if the CPU number is of type 'bool'. This is due
to 'bool' being a sub-class of 'int'. In most cases they are interchangeable,
but not when used inside a f-string. Some examples:
	f"{True}" 			-> "True"
	f"{True:d}" 			-> "1"
	{True : None}			-> {1 : None}
	1 in (False, True)		-> True

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>