_PepcPrinter: restructure printing code
Make the printing code less dependent on the key order in the 'csinfo' C-states
information dictionary.

Today, it just happen to have the "disable" key going first, by chance, so
printing works as expected, and we first print the C-state on/off status.

However, in the following patch we'll add the "desc" key printing, which goes
before the "disable" key. And with current code-base, we'll end up with
something like this:

    - description: MWAIT 0x00
C1: 'on' for CPUs 0-15 (all CPUs)
    - expected latency: 2 us
    - target residency: 2 us

Restructure the code so that the on/off status is always printed first.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>