MSR: continue without transaction when caching is disabled
We can either write the value each time to MSR or add it for transaction, too
be written later. For it to be possible to add to transaction,
'self._in_transaction' has to be 'True'. The only way to set it to 'True'
is to call 'start_transaction()'.

Thus, when caching is disabled 'start_transaction()' should print debug message
that it will not work and not change 'self._in_transation' to 'True'.
Because of 'self._in_transation' is not 'True', '_add_for_transation()' is
not called (if it is, it will raise an exception). And 'commit_transaction()'
will not do anything if caching is disabled. In conclusion, everything
will be work but without using transaction.

TL;DR: instead 'start_transaction()' raising an exception when caching is
disabled, it will continue without transaction, writing each time.

Signed-off-by: Niklas Neronin <niklas.neronin@intel.com>